+++ src/fava/templates/_layout.html
@@ -33,7 +33,7 @@
         </div>{% endif %}
       </h1>
     </header>
-    <button is="aside-button" class="aside-button" type="button">
+    <button id="aside-button" class="aside-button" type="button">
       <svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink" height="32px" width="32px" version="1.0" viewBox="0 0 24 24"><circle cx="12" cy="12" r="2"/><circle cx="12" cy="5" r="2"/><circle cx="12" cy="19" r="2"/></svg>
     </button>
     <aside>{% include "_aside.html" %}</aside>
